+++ src/modules/generator.js
@@ -20,28 +20,6 @@ module.exports = class KareeGenerator{
         }else if(this.settings.isScreen){
             this.generateScreen()
         }
-        // let spinner = new Spinner()
-        // spinner.setSpinnerString('⠋⠙⠹⠸⠼⠴⠦⠧⠇⠏')
-        // spinner.setSpinnerTitle('\x1b[32m\x1b[1mDownloading Flutter files\x1b[22m\x1b[0m')
-        // spinner.start()
-        // let res = cmd.run(`sleep 10`)
-        // res.on('close', (status, signal)=>{
-            
-        //     if(status == 0){
-
-        //         spinner.stop(false)
-
-        //         /**
-        //          * Ici on se déplace dans le projet flutter et on mets à jour la configuration
-        //          * dans le pubspec.yaml
-        //          */
-        //         console.log('\n')
-        //         console.log(options)
-        //     }else {
-        //         options.callback?.call(1);
-        //     }
-        // });
-
     }
 
     generateController(){
@@ -63,33 +41,12 @@ module.exports = class KareeGenerator{
         io.createDir('lib/app/controllers/'+(this.settings.path == '/' ? '' : this.settings.path))
         io.writeFile(template, `lib/app/controllers/${(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_controller.dart`)
 
-        /**
-         * Update now the list of controller
-         */
-        let controllersRegistered = io
-            .readFile('lib/app/controllers.dart', false)
-            .toString()
-            
-        controllersRegistered = controllersRegistered
-                .substring(0, controllersRegistered.indexOf(']'))
-                .trim()
-        
-        if( controllersRegistered.indexOf(formatter.underscoreToCambel(this.settings.className)+'Controller()') >= 0){
-            console.log(`\n\x1b[36m${this.settings.className}Controller\x1b[0m generated in lib/app/controllers/${(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_controller.dart\n`)
-            return;
-        }
-
-        controllersRegistered = `import "package:${this.projectConfig.appName}/app/controllers/${(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_controller.dart";\n`+
-        controllersRegistered+
-        `\n\t${this.settings.className}Controller(),\n];`
-        
-        io.writeFile(controllersRegistered, 'lib/app/controllers.dart')
-
         console.log(`\n\x1b[36m${this.settings.className}Controller\x1b[0m generated in lib/app/controllers/${(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_controller.dart\n`)
 
     }
 
     generateScreen(){
+        console.log(this.settings)
         validator.validateGeneratedScreen(this.settings)
         let template = ''
         if(this.settings.isStatefull)
@@ -102,8 +59,15 @@ module.exports = class KareeGenerator{
             .replace('$className', formatter.underscoreToCambel(this.settings.className))
             .replace('$className', formatter.underscoreToCambel(this.settings.className))
             .replace('$className', formatter.underscoreToCambel(this.settings.className))
-            .replace('$name', formatter.cambelToUnderscore(this.settings.name))
-            .replace('$name', formatter.cambelToUnderscore(this.settings.name))
+        if(this.settings.name === null || this.settings.name === undefined){
+            template = template .replace('@Screen("$name")', formatter.cambelToUnderscore(''))
+            .replace('$name', formatter.cambelToUnderscore(''))
+            console.log(template)
+        }else{
+            template = template
+                .replace('$name', formatter.cambelToUnderscore(this.settings.name))
+                .replace('$name', formatter.cambelToUnderscore(this.settings.name))
+        }
 
         if(this.settings.isStatefull)
             template = template
@@ -114,7 +78,6 @@ module.exports = class KareeGenerator{
          * On Crée l'arborescence indiqué par l'option --path où ajouter l'écran
          */
         io.createDir('lib/app/screens/'+(this.settings.path == '/' ? '' : this.settings.path))
-        // console.log(`\n#######################################\n## PATH = ${io.createDir('lib/app/screens/'+(this.settings.path == '/' ? '' : this.settings.path))}${formatter.cambelToUnderscore(this.settings.className)}_screen.dart\n##############################`)
         io.writeFile(template, `${'lib/app/screens/'+(this.settings.path == '/' ? '' : this.settings.path)}${formatter.cambelToUnderscore(this.settings.className)}_screen.dart`)
 
         /**