+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 # Release Notes
 
+## V1.3.4
+ * Update: Update installer! update core file to 1.0.4
+## V1.3.3
+ * Update: Update installer! remove generated core file from new project
+## V1.3.2
+ * Update: Update installer! remove generated file at this step and README
+## V1.3.1
+ * Update Adding config file yaml Karee v2.1.0
+## V1.3.0
+ * Update To point out to Karee v2.1.0
+## V1.2.0
+ * Update Dart Sdk version
 ## V1.0.0
  * First version
+## V1.0.1
+ * Applying Karee core files from remote